Change COGS connection to use EventTarget. Pave the way for Audio Player in this lib
