+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@clockworkdog/cogs-client",
-  "version": "0.10.0",
+  "version": "0.11.0",
   "main": "dist/index.js",
   "unpkg": "dist/browser/index.js",
   "scripts": {
@@ -15,6 +15,7 @@
   "author": "Clockwork Dog <info@clockwork.dog>",
   "license": "MIT",
   "devDependencies": {
+    "@types/howler": "^2.2.2",
     "@typescript-eslint/eslint-plugin": "^4.12.0",
     "@typescript-eslint/parser": "^4.12.0",
     "browserify": "^17.0.0",
@@ -25,6 +26,7 @@
     "typescript": "^4.1.3"
   },
   "dependencies": {
+    "howler": "^2.2.1",
     "reconnecting-websocket": "^4.4.0"
   },
   "description": "Connect to COGS to build a custom Media Master",
