+++ src/AudioPlayer.ts
@@ -1,0 +1,12 @@
+// import { Howl, Howler } from 'howler';
+// import { AudioClip } from './types/AudioState';
+
+// interface InternalClipPlayer extends AudioClip {
+//   player: Howl;
+// }
+
+// export default class AudioPlayer {
+//   constructor(connection: CogsConnection) {
+//     // connection.on('message')
+//   }
+// }
