+++ src/index.ts
@@ -1,4 +1,4 @@
-export { default as createCogsConnection } from './createCogsConnection';
+export { default as CogsConnection } from './CogsConnection';
 export { default as CogsClientMessage } from './types/CogsClientMessage';
 export { default as Callbacks } from './types/Callbacks';
 export { default as MediaObjectFit } from './types/MediaObjectFit';
